% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{actc_2017}
\alias{actc_2017}
\title{Australian Customs Tariff Classification (ACTC)}
\format{Data set \code{actc_2017} contains the following variables:
\describe{
  \item{section_number}{ACTC Schedule 3 section number}
  \item{section_name}{ACTC Schedule 3 section name}
  \item{chapter_number}{ACTC Schedule 3 chapter number}
  \item{chapter_name}{ACTC Schedule 3 chapter name}
  \item{heading_number_4digit}{ACTC Schedule 3 heading number}
  \item{heading_name_4digit}{ACTC Schedule 3 heading name}
  \item{subheading_number_5digit}{ACTC Schedule 3 subheading number (five-digit)}
  \item{subheading_name_5digit}{ACTC Schedule 3 subheading name (five-digit)}
  \item{subheading_number_6digit}{ACTC Schedule 3 subheading number (six-digit)}
  \item{subheading_name_6digit}{ACTC Schedule 3 subheading name (six-digit)}
  \item{reference_number}{ACTC eight-digit commodity reference number}
  \item{statistical_code}{ACTC Schedule 3 statistical code}
  \item{unit}{unit of measure}
  \item{commodity_name}{ACTC eight-digit commodity reference number}
  \item{rate}{current customs tariff rate}
  \item{notes}{notes to customs tariff}
}}
\source{
The ACTC is published and maintained by the Australian Government Department of Immigration and Border Protection.
}
\usage{
actc_2017
}
\description{
A data set containing the Australian Customs Tariff Classification (ACTC).
}
\details{
The Combined Australian Customs Tariff Nomenclature and Statistical
  Classification (commonly known as the Working Tariff) incorporates the
  \emph{Customs Tariff Act 1995} (Act No. 147 of 1995), as amended or proposed to be altered
  from time to time, and the Australian Harmonized Statistical Classification.

  The Australian Customs Tariff Classification working document is designed for day to day
  use in determining tariff classifications and rates of Customs duty, etc., for the
  purposes of entry of imported goods into Australia. 

  The Australian Customs Tariff Classification is based on the International Convention
  on the Harmonized Commodity Description and Coding System (HS) to which Australia is a
  signatory.  The Convention provides the Interpretation Rules (set out in Schedule 2),
  Sections, Chapters, headings, five and six digit subheadings and Notes to Sections and
  Chapters are contained in Schedule 3 of the
  \href{https://www.border.gov.au/Tariffclassificationofgoods/Documents/httitlgwv2.pdf}{Combined Australian Customs Tariff Nomenclature and Statistical Classification}.
 
  This dataset is based on the online version of Schedule 3 of the Department of Immigration
  and Border Protection document titled \href{https://www.border.gov.au/Busi/cargo-support-trade-and-goods/importing-goods/tariff-classification-of-goods/current-tariff-classification/schedule-3}{Combined Australian Customs Tariff Nomenclature
  and Statistical Classification}.

  This version of the Working Tariff contains the tariff changes resulting from the
  World Customs Organization fourth review of the International Convention on the
  Harmonized Commodity Description and Coding System, commonly referred to as the
  Harmonized System.  It also contains the consequential changes to the statistical
  codes in Schedule 3 and the statistical code changes made by the Australian Bureau
  of Statistics as a result of its review of the Statistical codes in Schedule 3 of
  the Tariff. The changes contained in the new Working Tariff commenced on 1 January 2017.

  For further details, see:
    \href{https://www.border.gov.au/Busi/cargo-support-trade-and-goods/importing-goods/tariff-classification-of-goods/current-tariff-classification}{Australian Government Department of Immigration and Border Protection, Current tariff classification}.
}
\references{
Department of Immigration and Border Protection 2017, \emph{Australian Harmonized Export Commodity Classification (ACTC), Schedule 3 - Classification of goods and general and special rates of duty}, DIBP, Canberra. URL: \url{https://www.border.gov.au/Busi/cargo-support-trade-and-goods/importing-goods/tariff-classification-of-goods/current-tariff-classification/schedule-3}.
}
\keyword{datasets}
